불필요 코드 정리
