+++ app/src/main/java/com/example/apriltagapp/model/baseModel/Pos.kt
@@ -3,5 +3,4 @@ package com.example.apriltagapp.model.baseModel
 abstract class Pos {
     /** 좌표계 차원 */
     abstract val dimension: Int
-
 }
